requirements.txt
# Core Django
Django==5.0.1
djangorestframework==3.14.0

# Database
psycopg2-binary==2.9.9

# Environment variables
python-decouple==3.8
python-dotenv==1.0.0

# Redis
redis==5.0.1
django-redis==5.4.0

# Image handling
Pillow==10.2.0

# CORS
django-cors-headers==4.3.1

# Authentication
djangorestframework-simplejwt==5.3.1

# Utilities
pytz==2023.3
python-dateutil==2.8.2

# Development
django-debug-toolbar==4.2.0
django-extensions==3.2.3

# Testing
pytest==7.4.3
pytest-django==4.7.0
coverage==7.3.4

# Code quality
black==23.12.1
flake8==7.0.0
isort==5.13.2

# API Documentation
drf-spectacular==0.27.0

# Celery (para tareas asincronas)
celery==5.3.4

# Validaciones
django-phonenumber-field==7.3.0
phonenumbers==8.13.27

# Seguridad
django-filter==23.5